[storage] Introduce `store`
## Overview

Introduces a new `store` module that contains a simple, unauthenticated `Store` K/V DB. This DB behaves identically to the `adb::Any` DB, but it does not maintain an MMR over the logs.

I didn't have enough time tonight to update the `adb` to use the new `Store` as a base, but a follow-up PR could de-duplicate some of the logic to compose the authenticated K/V DBs from the base, unauthenticated `Store`.

closes https://github.com/commonwarexyz/monorepo/issues/1206