WIP: Attempt to fix memory leak in deterministic runtime
This attempt makes it so that the deterministic context only has a weak
reference to the executor. The idea is that since the executor owns
tasks, which can own contexts, which can own the executor, you have
an issue of circular dependency. To break this, we can make it so that
the contexts don't cause the executor to be held on to.

Unfortunately, it doesn't seem like this has solved the leak completely.
Valgrind still detects about 20 MB of leakage from a single run.