[stream] link listener response to dialer hello
Closes #1481.

This binds the signed response from the listener to the hello message
received from the dialer.

This is implemented by augmenting the Info message with a "continuation tag".
Info has a new method to add a tag, computed as a hash over some data.
This tag, if present, will then get included in the signature in the hello
message.

This could be made even more fool proof by integrating the proposed transcript
construction, but I expect that will be better done with a refactor
of the stream crate to separate out the cryptographic AKE into a proper
abstraction.