[codec] Add impl for ()
This might seem a bit pointless, but this can show up if you have some
kind of generic abstraction requiring codec, but want to use () for that type.

For example, in the coding abstraction work, I want a type in the abstraction
to implement Codec, but in some impls, that type is (), so this impl is useful.