[coding] Add an extra check step before decoding
Closes #1686

This implements an extra step, wherein after receiving the reshards
from other participants, you're forced to check them individually,
before passing them to decode. The intention is that if you receive
a reshard, and it fails this check, then you don't even consider it
as progress towards the minimum number of shards you need to reconstruct
the data.

This avoids issues where someone can cause reconstruction to fail
by sending an invalid shard.