[consensus] `Application` interface
Adds a higher-level abstraction for building consensus applications with marshal. This interface enables building layers between `Automaton` and an `Application`'s core logic, which can facilitate many different patterns such as epoching, erasure coding, etc. to be implemented as reusable components.

As the first use of this interface, a new `EpochedApplication` layer is provided in `consensus::application`, handling:
1. Fetching "genesis" blocks for epochs > 0.
1. Fetching the parent block to build upon in the proposal path.
1. Re-proposing blocks on the epoch boundary.
1. Verifying re-proposed blocks on the epoch boundary.

This layer is used in the `commonware-reshare` example to showcase an application of the new interface.