[cryptography] Make polynomial codec accept maximum size
Instead of accepting an exact size, we accept a maximum size.

At this low a level of abstraction, being this strict about the size
is not good.

In the context of a DKG, yes, you do indeed want to check that the polynomial
has the right size, exactly, but this is done by application code.

Having an exact size forces you to leak an implementation detail all around
your codebase, because many parts now need to be aware of what the exact
size of a polynomial is, instead of just providing an upper bound.

In general, I think that decoding can try and prevent resource exhaustion,
but should not be responsible for validity in general.