[coding] Avoid storing more data than necessary in fuzz
We were previously storing the checking data and reshard
for every participant. Now we just storing all the shards
(which can share memory), one checking data + reshard,
and then only materialize the final checked shards,
avoiding storing the intermediates for the other players

This shrinks memory usage down a lot for ZODA,
because the checksum matrix is pretty large,
and we were storing one copy of it for every participant.

Why is it so large? There seems to be an issue there,
that we should fix subsequently.