[cryptography] Remove From<u32> and From<u64> on Scalar
These were a bit of a footgun in that they temptingly provided the same
API as Scalar::from_index, but subtly doing the wrong thing.

Thankfully, they weren't actually used throughout the codebase, so we
can remove them for now.

I don't really envision a scenario where we'd need these. In most cases,
you're generating a random scalar or decoding it.