[p2p] `CheckedSender::Error: 'static`
Fixes an issue where `CheckedSender::Error` was inferred to have the same lifetime as `LimitedSender::Checked<'_>` when using the `Sender` trait. In `main`, this makes it a pain to bubble up `CheckedSender::Error` with `?`, since the compiler infers the lifetime of the error to be tied to `CheckedSender`. In practice, it's unlikely that we'll want to return errors containing borrowed data.